commit fix model

+++ AppMusic/app/src/main/java/com/example/appmusic/data/model/DataPlayList.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 @Entity
 data class DataPlayList(
         @PrimaryKey(autoGenerate = true)
-        var idPlayList: Long = 0,
+        var idPlayList: Int = 0,
         @ColumnInfo
         var namePlayList: String = "",
         @ColumnInfo
